+++ core-build/gulp-core-build/src/tasks/JestTask.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 import * as path from 'path';
-import * as fsx from 'fs-extra';
 import { GulpTask} from './GulpTask';
 import { IBuildConfig } from '../IBuildConfig';
 import * as Gulp from 'gulp';
 import * as Jest from 'jest-cli';
 import * as globby from 'globby';
+import { FileSystem } from '@microsoft/node-core-library';
 
 /**
  * Configuration for JestTask
@@ -78,7 +78,7 @@ const DEFAULT_JEST_CONFIG_FILE_NAME: string = 'jest.config.json';
  */
 export function _isJestEnabled(rootFolder: string): boolean {
   const taskConfigFile: string = path.join(rootFolder, 'config', 'jest.json');
-  if (!fsx.existsSync(taskConfigFile)) {
+  if (!FileSystem.exists(taskConfigFile)) {
     return false;
   }
   const taskConfig: {} = require(taskConfigFile);
@@ -130,7 +130,7 @@ export class JestTask extends GulpTask<IJestConfig> {
     const jestConfig: any = {
       ci: this.buildConfig.production,
       cache: !!this.taskConfig.cache,
-      config: fsx.existsSync(configFileFullPath) ? configFileFullPath : undefined,
+      config: FileSystem.exists(configFileFullPath) ? configFileFullPath : undefined,
       collectCoverageFrom: this.taskConfig.collectCoverageFrom,
       coverage: this.taskConfig.coverage,
       coverageReporters: this.taskConfig.coverageReporters,
@@ -204,8 +204,8 @@ export class JestTask extends GulpTask<IJestConfig> {
     const snapDestFile: string = destinationFile.replace(/\.test\..+\.snap$/, extension);
     const testFileName: string = path.basename(snapDestFile, '.snap');
     const testFile: string = path.resolve(path.dirname(snapDestFile), '..', testFileName); // Up from `__snapshots__`.
-    if (fsx.existsSync(testFile)) {
-      fsx.copySync(snapSourceFile, snapDestFile);
+    if (FileSystem.exists(testFile)) {
+      FileSystem.copyFile(snapSourceFile, snapDestFile);
       return true;
     } else {
       return false;