+++ core-build/web-library-build/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log - @microsoft/web-library-build
 
-This log was last generated on Sat, 23 Jun 2018 02:21:20 GMT and should not be manually modified.
+This log was last generated on Tue, 03 Jul 2018 21:03:31 GMT and should not be manually modified.
+
+## 4.4.54
+Tue, 03 Jul 2018 21:03:31 GMT
+
+*Version update only*
+
+## 4.4.53
+Fri, 29 Jun 2018 02:56:51 GMT
+
+*Version update only*
 
 ## 4.4.52
 Sat, 23 Jun 2018 02:21:20 GMT