+++ webpack/loader-raw-script/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log - @microsoft/loader-raw-script
 
-This log was last generated on Sat, 23 Jun 2018 02:21:20 GMT and should not be manually modified.
+This log was last generated on Tue, 03 Jul 2018 21:03:31 GMT and should not be manually modified.
+
+## 1.2.48
+Tue, 03 Jul 2018 21:03:31 GMT
+
+*Version update only*
+
+## 1.2.47
+Fri, 29 Jun 2018 02:56:51 GMT
+
+*Version update only*
 
 ## 1.2.46
 Sat, 23 Jun 2018 02:21:20 GMT